+++ legacy/D3Launch/eax.h
@@ -48,7 +48,7 @@ typedef enum
 // use this structure for get/set all properties...
 typedef struct 
 {
-    unsigned long environment;          // 0 to EAX_ENVIRONMENT_COUNT-1
+    uint32_t environment;          // 0 to EAX_ENVIRONMENT_COUNT-1
     float fVolume;                      // 0 to 1
     float fDecayTime_sec;               // seconds, 0.1 to 100
     float fDamping;                     // 0 to 1
